+++ src/Sulu/Bundle/ContactBundle/Tests/Unit/ContactSelectionContentTypeTest.php
@@ -22,6 +22,7 @@ use Sulu\Bundle\ContactBundle\Contact\ContactManagerInterface;
 use Sulu\Bundle\ContactBundle\Content\Types\ContactSelectionContentType;
 use Sulu\Bundle\ContactBundle\Util\CustomerIdConverter;
 use Sulu\Bundle\ContactBundle\Util\IndexComparator;
+use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStoreInterface;
 use Sulu\Component\Content\Compat\PropertyInterface;
 use Sulu\Component\Content\Compat\PropertyParameter;
 use Sulu\Component\Content\Compat\StructureInterface;
@@ -83,6 +84,16 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
      */
     private $serializer;
 
+    /**
+     * @var ReferenceStoreInterface
+     */
+    private $accountReferenceStore;
+
+    /**
+     * @var ReferenceStoreInterface
+     */
+    private $contactReferenceStore;
+
     protected function setUp()
     {
         parent::setUp();
@@ -99,6 +110,8 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
         $this->property->getStructure()->willReturn($this->structure->reveal());
 
         $this->serializer = $this->prophesize(Serializer::class);
+        $this->accountReferenceStore = $this->prophesize(ReferenceStoreInterface::class);
+        $this->contactReferenceStore = $this->prophesize(ReferenceStoreInterface::class);
     }
 
     public function testGetTemplate()
@@ -109,7 +122,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->assertEquals($this->template, $type->getTemplate());
@@ -123,7 +138,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -148,7 +165,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -173,7 +192,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -198,7 +219,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -223,7 +246,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -249,7 +274,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $nodeProperty = $this->prophesize(\PHPCR\PropertyInterface::class);
@@ -280,7 +307,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $view = $type->getViewData($this->property->reveal());
@@ -296,7 +325,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $defaultValue = $type->getDefaultValue();
@@ -312,7 +343,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $defaultParams = $type->getDefaultParams();
@@ -334,7 +367,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getName()->willReturn('test');
@@ -359,7 +394,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $contact1 = $this->prophesize(Contact::class);
@@ -400,7 +437,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $entity1 = $this->prophesize(Account::class);
@@ -441,7 +480,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $contact1 = $this->prophesize(Contact::class);
@@ -483,7 +524,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getValue()->willReturn([]);
@@ -503,7 +546,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getValue()->willReturn(null);
@@ -523,7 +568,9 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
             $this->accountManager->reveal(),
             $this->serializer->reveal(),
             new CustomerIdConverter(),
-            new IndexComparator()
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
         );
 
         $this->property->getValue()->willReturn('blabla');
@@ -533,5 +580,26 @@ class ContactSelectionContentTypeTest extends \PHPUnit_Framework_TestCase
         $result = $type->getContentData($this->property->reveal());
 
         $this->assertCount(0, $result);
+    }
+
+    public function testPreResolve()
+    {
+        $type = new ContactSelectionContentType(
+            $this->template,
+            $this->contactManager->reveal(),
+            $this->accountManager->reveal(),
+            $this->serializer->reveal(),
+            new CustomerIdConverter(),
+            new IndexComparator(),
+            $this->accountReferenceStore->reveal(),
+            $this->contactReferenceStore->reveal()
+        );
+
+        $this->property->getValue()->willReturn(['a1', 'c1', 'a3']);
+        $type->preResolve($this->property->reveal());
+
+        $this->accountReferenceStore->add(1)->shouldBeCalled();
+        $this->accountReferenceStore->add(3)->shouldBeCalled();
+        $this->contactReferenceStore->add(1)->shouldBeCalled();
     }
 }