+++ src/Sulu/Bundle/MediaBundle/Tests/Unit/Media/ImageConverter/Transformations/PasteTransformationTest.php
@@ -11,6 +11,7 @@
 
 namespace Sulu\Bundle\MediaBundle\Tests\Unit\Media\ImageConverter\Transformations;
 
+use Imagine\Gd\Imagine;
 use Imagine\Image\Box;
 use Imagine\Image\ImageInterface;
 use Prophecy\Argument;
@@ -41,7 +42,8 @@ class PasteTransformationTest extends SuluTestCase
             __DIR__ . '/../../../../app/Resources/images/photo.jpeg'
         );
 
-        $this->pasteTransformation = new pasteTransformation(
+        $this->pasteTransformation = new PasteTransformation(
+            new Imagine(),
             $this->fileLocator->reveal()
         );
 
@@ -67,14 +69,12 @@ class PasteTransformationTest extends SuluTestCase
     public function testNoPaste()
     {
         $image = $this->prophesize(ImageInterface::class);
-        $image->getSize()->willReturn(new Box(700, 500));
-        $image->paste(Argument::any(), Argument::any())->shouldNotBeCalled();
+
+        $this->setExpectedException(\RuntimeException::class);
 
         $returnImage = $this->pasteTransformation->execute(
             $image->reveal(),
             []
         );
-
-        $this->assertInstanceOf(ImageInterface::class, $returnImage);
     }
 }