+++ src/Sulu/Component/Contact/SmartContent/AccountDataProvider.php
@@ -12,6 +12,7 @@
 namespace Sulu\Component\Contact\SmartContent;
 
 use JMS\Serializer\SerializerInterface;
+use Sulu\Bundle\WebsiteBundle\ReferenceStore\ReferenceStoreInterface;
 use Sulu\Component\SmartContent\Orm\BaseDataProvider;
 use Sulu\Component\SmartContent\Orm\DataProviderRepositoryInterface;
 
@@ -20,9 +21,9 @@ use Sulu\Component\SmartContent\Orm\DataProviderRepositoryInterface;
  */
 class AccountDataProvider extends BaseDataProvider
 {
-    public function __construct(DataProviderRepositoryInterface $repository, SerializerInterface $serializer)
+    public function __construct(DataProviderRepositoryInterface $repository, SerializerInterface $serializer, ReferenceStoreInterface $referenceStore)
     {
-        parent::__construct($repository, $serializer);
+        parent::__construct($repository, $serializer, $referenceStore);
 
         $this->configuration = self::createConfigurationBuilder()
             ->enableTags()