+++ src/Sulu/Component/HttpCache/Tests/Unit/Handler/TagsHandlerTest.php
@@ -85,6 +85,35 @@ class TagsHandlerTest extends \PHPUnit_Framework_TestCase
         $this->handler->flush();
     }
 
+    public function testInvalidateReference()
+    {
+        $this->proxyCache->ban(
+            [
+                TagsHandler::TAGS_HEADER => '(' . preg_quote('test-1') . ')(,.+)?$',
+            ]
+        )->shouldBeCalled();
+        $this->proxyCache->flush()->shouldBeCalled();
+
+        $this->handler->invalidateReference('test', 1);
+
+        $this->handler->flush();
+    }
+
+    public function testInvalidateReferenceDuplicated()
+    {
+        $this->proxyCache->ban(
+            [
+                TagsHandler::TAGS_HEADER => '(' . preg_quote('test-1') . ')(,.+)?$',
+            ]
+        )->shouldBeCalledTimes(1);
+        $this->proxyCache->flush()->shouldBeCalled();
+
+        $this->handler->invalidateReference('test', 1);
+        $this->handler->invalidateReference('test', 1);
+
+        $this->handler->flush();
+    }
+
     public function testUpdateResponse()
     {
         $id = Uuid::uuid4()->toString();